requirements.txt
streamlit
pymupdf
pytesseract
Pillow
langchain
langchain-text-splitters
fastembed
langchain-community
chromadb
langchain-chroma
python-dotenv
langchain-google-genai
